backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
sentence-transformers==2.2.2
chromadb==0.4.15
tiktoken==0.5.1
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.30.2
torch==2.0.1
huggingface-hub==0.16.4
tokenizers==0.13.3
Pillow==10.0.1
psutil==5.9.5